firewallbackend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
Pillow>=10.1.0
requests>=2.31.0
paramiko>=3.4.0
cryptography>=41.0.0
pyOpenSSL>=24.0.0
django-extensions>=3.2.3
django-csp==3.7
werkzeug>=3.0.1
gunicorn==21.2.0
whitenoise==6.6.0
ipaddress==1.0.23
pythonping==1.1.4
waitress==2.1.2
xlsxwriter==3.1.9
setuptools==65.6.3
jinja2==3.1.2
pyinstaller==6.3.0
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
uvicorn==0.27.1
uvloop==0.19.0
httptools==0.6.1
channels==4.0.0
daphne==4.0.0
argon2-cffi==23.1.0
pandas==2.1.4
celery>=5.3.0
redis>=5.0.0
